+++ R/PipeOpTorchFTCLS.R
@@ -65,7 +65,10 @@ inherit = PipeOpTorch,
         id = id,
         module_generator = nn_ft_cls,
         param_vals = param_vals,
-        param_set = param_set
+        param_set = param_set,
+        # only the token dimension is needed to build the module, the number of tokens
+        # may be unknown
+        only_batch_unknown = FALSE
       )
     }
   ),
@@ -74,6 +77,8 @@ inherit = PipeOpTorch,
       if (length(shapes_in$input) != 3) {
         stop("Input tensor must have 3 dimensions.")
       }
+      assert_known_dims(shapes_in[[1L]], 3L, "the token dimension (dimension 3)", self$id)
+      # if the number of tokens is unknown, adding the CLS token keeps it unknown
       shapes_in[[1]][2] = shapes_in[[1]][2] + 1
       return(shapes_in)
     },